Move neutral compute helpers into core/calc
