+++ src/core/storage/utest_data_merger.cpp
@@ -9,7 +9,7 @@
 #include <fstream>
 #include <unistd.h>
 #include <vector>
-#include "core/compute/compute.h"
+#include "core/calc/norm_utils.h"
 #include "core/storage/data_file.h"
 #include "core/storage/data_file_layout.h"
 #include "core/storage/data_merger.h"
